Refactored approve and active buttons
